# [async] Switch to llvm::SmallVector<llvm::SmallSet> for edges